requirements.txt
flask
flask_sqlalchemy
flask_login
flask_wtf
werkzeug
bcrypt
python-dotenv
gunicorn
psycopg2-binary
